convert nav menu to use vuejs
this change wraps the entire contents in a vue app and converts the nav
menu to use vue for switching the main content. this lightens the
developer toil when adding new sections as now there is only one main
place to add the button, and then one place to add the data.

it's not clear if this change has sped up the rendering.