refactor server
This change refactors the server command to take a url query parameter
which specifies a must-gather tar to download. It also removes the
previous functionality of the server taking a path parameter. It does
not affect the non-server functionality.

Also updates dockferfile, readme, and changelog.

usage is as follows:

```
$ okd-camgi --server
```

open web browser to
`http://localhost:8080/?url=http://path/to/must-gather.tar.gz`